Edited images for cleaner flow in animation
Adjusted positioning of images on background so that they would match up better (not as much shift in direction from one image to the other). The toDo img was initially further left than toDoStrikethru img and I could still see the top of the tick mark on the toDoStrikethru img even when top image was fully opaque.